fix: move metadata to profileHeader
